fs: initfs: Fix start_pos type

+++ src/fs/driver/initfs/initfs.h
@@ -15,23 +15,23 @@ struct inode;
 struct cpio_entry;
 
 struct initfs_file_info {
-	unsigned int    start_pos;
+	uintptr_t start_pos;
 	struct cpio_entry *entry;
 
 	/* Following fields are used just for directories */
-	char  *path;
+	char *path;
 	size_t path_len;
-	char  *name;
+	char *name;
 	size_t name_len;
 };
 
 extern struct initfs_file_info *initfs_alloc_inode(void);
 extern void initfs_free_inode(struct initfs_file_info *fi);
 extern int initfs_iterate(struct inode *next, char *name, struct inode *parent,
-			struct dir_ctx *ctx);
+    struct dir_ctx *ctx);
 extern int initfs_create(struct inode *i_new, struct inode *i_dir, int mode);
 extern int initfs_fill_inode(struct inode *node, char *cpio,
-			struct cpio_entry *entry);
+    struct cpio_entry *entry);
 
 extern int initfs_destroy_inode(struct inode *inode);
